S P I N N I N G B O O T
I'm going mad

+++ Assets/_TeamB_P02_AR/Scripts/ScriptableObjects/BaseCatchable.cs
@@ -8,6 +8,6 @@ public class BaseCatchable : ScriptableObject
     public string CatchableName;
     public GameObject CatchableObj;
     public bool IsJunk;
-    public float catchChance;
+    public double catchChance;
 }
